arrglamos los permisos

+++ app/Http/Controllers/PermissionController.php
@@ -22,7 +22,7 @@ class PermissionController extends Controller
     }}*/
     public function index()
     {
-        //$this->authorize('index',Role::class);
+        $this->authorize('index',Role::class);
         return view('backoffice.pages.permission.index',[
             'permissions'=>Permission::all(),
         ]);
@@ -35,7 +35,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
-        //$this->authorize('create',Role::class);
+        $this->authorize('create',Role::class);
         return view('backoffice.pages.permission.create',[
             'roles'=>Role::all(),
         ]);
@@ -62,7 +62,7 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
-        //$this->authorize('view', $permission);
+        $this->authorize('view', $permission);
         return view('backoffice.pages.permission.show',[
             'permission' => $permission
         ]);
@@ -76,7 +76,7 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
-        //$this->authorize('update', $permission);
+        $this->authorize('update', $permission);
         return view('backoffice.pages.permission.edit',[
             'permission'=>$permission,
             'roles'=>Role::all()
@@ -103,7 +103,7 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
-        //$this->authorize('delete', $permission);
+        $this->authorize('delete', $permission);
         $role = $permission->role;
         $permission->delete();
         return redirect()->route('backoffice.role.show',$role);
